Backport changes from v0.0.0-alpha.4

+++ pkg/svs/missing_data.go
@@ -23,17 +23,17 @@ package svs
 
 type MissingData interface {
 	Source() string
-	LowSeqno() uint
-	HighSeqno() uint
+	LowSeqno() uint64
+	HighSeqno() uint64
 }
 
 type missingData struct {
 	source    string
-	lowSeqno  uint
-	highSeqno uint
+	lowSeqno  uint64
+	highSeqno uint64
 }
 
-func NewMissingData(source string, low uint, high uint) MissingData {
+func NewMissingData(source string, low uint64, high uint64) MissingData {
 	return missingData{source: source, lowSeqno: low, highSeqno: high}
 }
 
@@ -41,10 +41,10 @@ func (md missingData) Source() string {
 	return md.source
 }
 
-func (md missingData) LowSeqno() uint {
+func (md missingData) LowSeqno() uint64 {
 	return md.lowSeqno
 }
 
-func (md missingData) HighSeqno() uint {
+func (md missingData) HighSeqno() uint64 {
 	return md.highSeqno
 }
