+++ test/svs/state_vector_test.go
@@ -31,14 +31,14 @@ import (
 
 func TestStateVectorBasics(t *testing.T) {
 	sv := svs.NewStateVector()
-	assert.Equal(t, uint(0), sv.Get("/node1"))
+	assert.Equal(t, uint64(0), sv.Get("/node1"))
 	sv.Set("/node1", 60)
 	sv.Set("/node2", 9)
 	sv.Set("/node1", 62)
 	sv.Set("/node3", 1)
-	assert.Equal(t, uint(62), sv.Get("/node1"))
-	assert.Equal(t, uint(9), sv.Get("/node2"))
-	assert.Equal(t, uint(1), sv.Get("/node3"))
+	assert.Equal(t, uint64(62), sv.Get("/node1"))
+	assert.Equal(t, uint64(9), sv.Get("/node2"))
+	assert.Equal(t, uint64(1), sv.Get("/node3"))
 	assert.Equal(t, int(3), sv.Len())
 }
 
@@ -60,8 +60,8 @@ func TestStateVectorEncodeDecode(t *testing.T) {
 	sv.Set("/two", 2)
 	comp := sv.ToComponent()
 	nsv, _ := svs.ParseStateVector(comp)
-	assert.Equal(t, uint(1), nsv.Get("/one"))
-	assert.Equal(t, uint(2), nsv.Get("/two"))
+	assert.Equal(t, uint64(1), nsv.Get("/one"))
+	assert.Equal(t, uint64(2), nsv.Get("/two"))
 	assert.Equal(t, int(2), nsv.Len())
 	assert.Equal(t, sv, nsv)
 }
@@ -69,7 +69,7 @@ func TestStateVectorEncodeDecode(t *testing.T) {
 func TestStateVectorDecodeStatic(t *testing.T) {
 	comp, _ := enc.ComponentFromBytes([]byte{201, 24, 202, 10, 7, 5, 8, 3, 111, 110, 101, 204, 1, 1, 202, 10, 7, 5, 8, 3, 116, 119, 111, 204, 1, 2})
 	sv, _ := svs.ParseStateVector(comp)
-	assert.Equal(t, uint(1), sv.Get("/one"))
-	assert.Equal(t, uint(2), sv.Get("/two"))
+	assert.Equal(t, uint64(1), sv.Get("/one"))
+	assert.Equal(t, uint64(2), sv.Get("/two"))
 	assert.Equal(t, int(2), sv.Len())
 }
